+++ docs/conf.py
@@ -3,19 +3,16 @@
 
 import os
 import sys
-import re
-sys.path.insert(0, "..")
-sys.path.insert(0, ".")
 
 
 # General information about the project.
 project = "OpenNMT-tf"
-copyright = "2017, The OpenNMT Authors"
+copyright = "2018, The OpenNMT Authors"
 author = "OpenNMT"
 language = "en"
 
-version = "1.8"  # The short X.Y version.
-release = "1.8.1"  # The full version, including alpha/beta/rc tags.
+version = "1.9"  # The short X.Y version.
+release = "1.9.0"  # The full version, including alpha/beta/rc tags.
 
 source_suffix = ".rst"
 master_doc = "index"
@@ -40,15 +37,16 @@ templates_path = ["_templates"]
 exclude_patterns = ["_build"]
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = "sphinx"
+pygments_style = "default"
 
 html_theme = "sphinx_rtd_theme"
 html_theme_options = {}
 html_static_path = ["_static"]
 html_show_sourcelink = False
-html_show_copyright = False
+html_show_copyright = True
 html_show_sphinx = False
 html_logo = "_static/logo-alpha.png"
+html_favicon = "_static/favicon.png"
 htmlhelp_basename = "opennmtdoc"
 html_sidebars = {
     "**": [
@@ -60,11 +58,6 @@ html_sidebars = {
 autodoc_member_order = "bysource"
 napoleon_include_init_with_doc = True
 napoleon_include_special_with_doc = True
-
-scv_whitelist_branches = ("none",)
-scv_whitelist_tags = (re.compile(r'^v\d+\.\d+\.0$'),)
-scv_sort = ("semver",)
-scv_greatest_tag = True
 
 def setup(app):
   app.add_stylesheet("custom.css")