+++ opennmt/models/model.py
@@ -7,10 +7,9 @@ import six
 
 import tensorflow as tf
 
-from opennmt.utils import data
+from opennmt.utils import data, hooks
 from opennmt.utils.optim import optimize
-from opennmt.utils.hooks import add_counter
-from opennmt.utils.misc import add_dict_to_collection, item_or_tuple
+from opennmt.utils.misc import item_or_tuple
 from opennmt.utils.parallel import GraphDispatcher
 
 
@@ -48,11 +47,14 @@ class Model(object):
     """
     return self._build(features, labels, params, mode, config=config)
 
-  def model_fn(self, num_devices=1):
+  def model_fn(self, num_devices=1, eval_prediction_hooks_fn=None):
     """Returns the model function.
 
     Args:
       num_devices: The number of devices used for training.
+      eval_prediction_hooks_fn: A callable that takes the model predictions
+        during evaluation and return an iterable of evaluation hooks (e.g. for
+        saving predictions on disk, running external evaluators, etc.).
 
     See Also:
       ``tf.estimator.Estimator`` 's ``model_fn`` argument for more details about
@@ -93,7 +95,11 @@ class Model(object):
     def _model_fn(features, labels, params, mode, config):
       """model_fn implementation."""
       if mode == tf.estimator.ModeKeys.TRAIN:
-        self._register_word_counters(features, labels)
+        counters = self._register_word_counters(features, labels)
+        counters_hook = hooks.CountersHook(
+            every_n_steps=config.save_summary_steps,
+            output_dir=config.model_dir,
+            counters=counters)
 
         features_shards = dispatcher.shard(features)
         labels_shards = dispatcher.shard(labels)
@@ -103,11 +109,12 @@ class Model(object):
               _loss_op, features_shards, labels_shards, params, mode, config)
 
         loss = _extract_loss(losses_shards)
-        train_op = optimize(loss, params)
+        train_op = optimize(loss, params, mixed_precision=(self.dtype == tf.float16))
         return tf.estimator.EstimatorSpec(
             mode,
             loss=loss,
-            train_op=train_op)
+            train_op=train_op,
+            training_hooks=[counters_hook])
       elif mode == tf.estimator.ModeKeys.EVAL:
         with tf.variable_scope(self.name):
           logits, predictions = self._build(features, labels, params, mode, config=config)
@@ -115,14 +122,14 @@ class Model(object):
 
         loss = _extract_loss(loss)
         eval_metric_ops = self._compute_metrics(features, labels, predictions)
-        if predictions is not None:
-          # Register predictions in a collection so that hooks can easily fetch them.
-          add_dict_to_collection("predictions", predictions)
-
+        evaluation_hooks = []
+        if predictions is not None and eval_prediction_hooks_fn is not None:
+          evaluation_hooks.extend(eval_prediction_hooks_fn(predictions))
         return tf.estimator.EstimatorSpec(
             mode,
             loss=loss,
-            eval_metric_ops=eval_metric_ops)
+            eval_metric_ops=eval_metric_ops,
+            evaluation_hooks=evaluation_hooks)
       elif mode == tf.estimator.ModeKeys.PREDICT:
         with tf.variable_scope(self.name):
           _, predictions = self._build(features, labels, params, mode, config=config)
@@ -204,11 +211,13 @@ class Model(object):
     features_length = self._get_features_length(features)
     labels_length = self._get_labels_length(labels)
 
+    counters = []
     with tf.variable_scope("words_per_sec"):
       if features_length is not None:
-        add_counter("features", tf.reduce_sum(features_length))
+        counters.append(hooks.add_counter("features", tf.reduce_sum(features_length)))
       if labels_length is not None:
-        add_counter("labels", tf.reduce_sum(labels_length))
+        counters.append(hooks.add_counter("labels", tf.reduce_sum(labels_length)))
+    return counters
 
   def _initialize(self, metadata):
     """Runs model specific initialization (e.g. vocabularies loading).