+++ opennmt/tokenizers/opennmt_tokenizer.py
@@ -1,6 +1,7 @@
 """Define the OpenNMT tokenizer."""
 
 import copy
+import six
 
 import tensorflow as tf
 
@@ -29,19 +30,22 @@ def create_tokenizer(config):
 class OpenNMTTokenizer(Tokenizer):
   """Uses the OpenNMT tokenizer."""
 
-  def __init__(self, configuration_file_or_key=None):
-    super(OpenNMTTokenizer, self).__init__(configuration_file_or_key=configuration_file_or_key)
-    self._tokenizer = None
+  def __init__(self, *arg, **kwargs):
+    super(OpenNMTTokenizer, self).__init__(*arg, **kwargs)
+    self._tokenizer = create_tokenizer(self._config)
+
+  def initialize(self, metadata):
+    super(OpenNMTTokenizer, self).initialize(metadata)
+    self._tokenizer = create_tokenizer(self._config)
+    for key, value in six.iteritems(self._config):
+      if key.endswith("path"):
+        tf.add_to_collection(tf.GraphKeys.ASSET_FILEPATHS, tf.constant(value))
 
   def _tokenize_string(self, text):
-    if self._tokenizer is None:
-      self._tokenizer = create_tokenizer(self._config)
     text = tf.compat.as_bytes(text)
     tokens, _ = self._tokenizer.tokenize(text)
     return [tf.compat.as_text(token) for token in tokens]
 
   def _detokenize_string(self, tokens):
-    if self._tokenizer is None:
-      self._tokenizer = create_tokenizer(self._config)
     tokens = [tf.compat.as_bytes(token) for token in tokens]
     return tf.compat.as_text(self._tokenizer.detokenize(tokens))